# api-docs/SwiftChat-api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/messages":{"post":{"tags":["Message"],"operationId":"saveMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRequest"}}},"required":true},"responses":{"201":{"description":"Created"}}},"patch":{"tags":["Message"],"operationId":"setMessagesToSeen","parameters":[{"name":"chatId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Accepted"}}}},"/api/v1/messages/upload-media":{"post":{"tags":["Message"],"operationId":"uploadMessage","parameters":[{"name":"chatId","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"201":{"description":"Created"}}}},"/api/v1/chats":{"get":{"tags":["Chat"],"operationId":"getChatsByRecipientId","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatResponse"}}}}}}},"post":{"tags":["Chat"],"operationId":"createChat","parameters":[{"name":"senderId","in":"query","required":true,"schema":{"type":"string"}},{"name":"recipientId","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StringResponse"}}}}}}},"/api/v1/users ":{"get":{"tags":["User"],"operationId":"getAllUsersExceptSelf","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}}},"/api/v1/messages/chat/{chatId}":{"get":{"tags":["Message"],"operationId":"getMessagesByChatId","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"}}}}}}}}},"components":{"schemas":{"MessageRequest":{"type":"object","properties":{"content":{"type":"string"},"senderId":{"type":"string"},"recipientId":{"type":"string"},"messageType":{"type":"string","enum":["TEXT","IMAGE","AUDIO","VIDEO"]},"chatId":{"type":"string"}}},"StringResponse":{"type":"object","properties":{"response":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"lastSeen":{"type":"string","format":"date-time"},"online":{"type":"boolean"}}},"MessageResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"state":{"type":"string","enum":["SENT","SEEN"]},"type":{"type":"string","enum":["TEXT","IMAGE","AUDIO","VIDEO"]},"senderId":{"type":"string"},"recipientId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"media":{"type":"string","format":"byte"}}},"ChatResponse":{"type":"object","properties":{"name":{"type":"string"},"unreadCount":{"type":"integer","format":"int64"},"lastMessage":{"type":"string"},"lastMessageTime":{"type":"string","format":"date-time"},"senderId":{"type":"string"},"recipientId":{"type":"string"},"id":{"type":"string"},"recipientOnline":{"type":"boolean"}}}},"securitySchemes":{"keycloak":{"type":"oauth2","in":"header","scheme":"bearer","bearerFormat":"JWT","flows":{"password":{"authorizationUrl":"http://localhost:9090/realms/SwiftChat/protocol/openid-connect/auth","tokenUrl":"http://localhost:9090/realms/SwiftChat/protocol/openid-connect/token","scopes":{}}}}}}}